refactor index.js

+++ bin/index.js
@@ -15,35 +15,41 @@ const dirName = options.dir;
 
 const directoryPath = path.join(dirName);
 
-const run = chalk.white.bold("Running.. ");
+const run = chalk.yellow.bold("Running.. ");
 const logFile = chalk.yellow.bold("Creating log file.. ");
 const done = chalk.white.bold("All done");
 const temp = chalk.yellow.bold('Creating temp folder..');
-const process = chalk.white.bold('Process closed');
+const tempExist = chalk.yellow.bold('temp folder exist');
+const isTempExist = chalk.yellow.bold('Checking if temp folder exist');
 
 console.log(run);
 
+console.log(isTempExist);
+if (fs.existsSync('./temp')) {
+    return console.log(tempExist);
+}
 
 const headers = ['File', 'Size', 'Conversion Time', '\n'].join('\t|\t');
 
-console.log(temp);
 fs.mkdir('./temp',(err) => {
     if(err){
-        console.log(chalk.red.bold("err: ", JSON.stringify(err)));
+        return console.log(chalk.red.bold("err: ", JSON.stringify(err)));
     }
+    console.log(temp);
 });
 
-console.log(logFile);
 fs.writeFile('./temp/results.txt', headers, (err)=> {
     if(err){
-        console.log(chalk.red.bold("err: ", err));
+        return console.log(chalk.red.bold("err: ", err));
     }
+    console.log(logFile);
 });
 
-console.log(chalk.yellow.bold('reading directory ' + directoryPath));
+// console.log(chalk.yellow.bold('reading directory ' + directoryPath));
 
 fs.readdir(directoryPath, function (err, files) {
     let start;
+    let time;
     //handling error
     if (err) {
         return console.log(chalk.red.bold('Unable to scan directory: ' + err));
@@ -51,37 +57,22 @@ fs.readdir(directoryPath, function (err, files) {
 
     //running ffmpeg on all files using forEach
     files.forEach((file, index) => {
+        // start = 0;
         // Create message per file
-        const create = chalk.yellow.bold("Converting file.. ", index);
-        const log = chalk.yellow.bold("Logging The results for file.. ", index);
-
-        const child = child_process.exec(`ffmpeg -stats -i ${directoryPath}\\${file} -max_muxing_queue_size 9999 ./temp/output${index}.mp4`, (err, stdout, stderr) => {
-            start = Date.now();
-            console.log(chalk.yellow.bold("Processing file.. " + index));
-            if (err) {
-                console.log(chalk.red.bold("err: ",err));
-                return;
+        start = Date.now();
+        child_process.execSync(`ffmpeg -i ${directoryPath}\\${file} -max_muxing_queue_size 9999 ./temp/output${index}.mp4`, {stdio: "ignore"});
+        time =  Math.floor(((Date.now() - start)/1000)*100) / 100;
+        const stats = fs.statSync(`${directoryPath}\\${file}`);
+        const fileSizeInBytes = stats.size;
+        const truncatedNumber = Math.floor((fileSizeInBytes*0.000001)*100) / 100;
+        const rowData = [path.extname(file),  truncatedNumber, time+'sec', '\n'].join('\t|\t');
+        fs.appendFile('./temp/results.txt', rowData, (err) =>{
+            if(err){
+                return console.log(chalk.red.bold("err: ", err));
             }
-            console.log(create);
-        });
-        child.on('exit', (d)=>{
-           console.log(chalk.red('exit'));
-        });
-        child.on('close', () => {
-            console.log(chalk.red('close'));
-            const time = Date.now() - start;
-            const stats = fs.statSync(`${directoryPath}\\${file}`);
-            const fileSizeInBytes = stats.size;
-            const truncatedNumber = Math.floor((fileSizeInBytes*0.000001)*100) /100;
-            const rowData = [path.extname(file),  truncatedNumber, time+'sec', '\n'].join('\t|\t');
+            const log = chalk.yellow.bold("Logging The results for file.. ", index);
             console.log(log);
-            fs.appendFile('./temp/results.txt', rowData, (err) =>{
-                if(err){
-                    console.log(chalk.red.bold("err: ", err));
-                }
-            });
             console.log(done);
-            console.log(process);
         });
     });
 });
